requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cycler==0.11.0
Django==3.2.25
djangorestframework==3.15.1
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
mysqlclient==2.1.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
zipp==3.6.0
